requirements.txt
boto3
openpyxl
pandas
anthropic
python-dotenv